Initial code push
Adds an audio filter named "Wave Link Sync" which has 2 dropdowns - the Channel and the Mixer.
When applied to an audio input or specifically audio output (optimally for the same Wave Link channel) it will apply volume adjustments based on what the percentage is set to in Wave Link itself (either for the monitor mix or stream mix)

Additionally, it supports un/muting the channel.

It will apply additional volume adjustments and mute status based on the main mixer volume (at the bottom of Wave Link)
There will probably be extra configuration options / checkboxes inside the filter.

The main use-case for this is to include the, for example, Voice Chat channel and be able to use it for sidechaining / ducking on the main Stream Mix.