Add new options to adhere to more Wave Link stuff
Namely checkboxes to follow the channel mute status (for the selected mixer), applying the main mixer volume and following the mixer mute status.

Additionally, the volume curve has been adjusted to fit much more closely to what it should be in Wave Link. Of course, keep in mind this is all before VSTs are applied, so if you have a VST applied that would affect the Gain on the audio, you would also need to put a Gain filter before this one in OBS.